requirements.txt
pip~=24.0
pathlib~=1.0.1
configparser~=6.0.0
distro~=1.9.0
wheel~=0.38.4
filelock~=3.13.1
redis~=5.0.1
pytz~=2024.1
setuptools~=65.5.1
Django~=4.2.10
asgiref~=3.7.2
sqlparse~=0.4.4
psycopg~=3.1.18
numpy~=1.26.4
requests~=2.31.0
packaging~=23.2
djangorestframework~=3.14.0
pyparsing~=3.1.1
fitz~=0.0.1.dev2
openai~=1.12.0
celery~=5.3.6
psycopg2-binary
PyMuPDF~=1.23.26
SQLAlchemy~=2.0.28